Extract project creation from example logic to new services files
