+++ newIDE/app/src/BrowserApp.js
@@ -1,14 +1,11 @@
 // @flow
 import * as React from 'react';
-import { t } from '@lingui/macro';
-import { type I18n as I18nType } from '@lingui/core';
 import MainFrame from './MainFrame';
 import Window from './Utils/Window';
 import ExportDialog from './Export/ExportDialog';
 import CreateProjectDialog from './ProjectCreation/CreateProjectDialog';
 import Authentication from './Utils/GDevelopServices/Authentication';
 import './UI/iconmoon-font.css'; // Styles for Iconmoon font.
-import { type ExampleShortHeader } from './Utils/GDevelopServices/Example';
 
 // Import for browser only IDE
 import BrowserExamples from './ProjectCreation/BrowserExamples';
@@ -30,43 +27,7 @@ import DownloadFileStorageProvider from './ProjectsStorage/DownloadFileStoragePr
 import DropboxStorageProvider from './ProjectsStorage/DropboxStorageProvider';
 import OneDriveStorageProvider from './ProjectsStorage/OneDriveStorageProvider';
 import { BrowserResourceFetcher } from './ProjectsStorage/ResourceFetcher/BrowserResourceFetcher';
-import { type StorageProvider, type FileMetadata } from './ProjectsStorage';
-
-// Import for project creation from example
-import { getExample } from './Utils/GDevelopServices/Example';
-import { sendNewGameCreated } from './Utils/Analytics/EventSender';
-import { showErrorBox } from './UI/Messages/MessageBox';
-
-const onCreateFromExampleShortHeader = (
-  isOpeningCallback: boolean => void,
-  onOpenCallback: (
-    storageProvider: StorageProvider,
-    fileMetadata: FileMetadata
-  ) => void
-) => async (
-  i18n: I18nType,
-  exampleShortHeader: ExampleShortHeader,
-) => {
-  try {
-    isOpeningCallback(true);
-    const example = await getExample(exampleShortHeader);
-    onOpenCallback(UrlStorageProvider, {
-      fileIdentifier: example.projectFileUrl,
-    });
-    sendNewGameCreated(example.projectFileUrl);
-  } catch (error) {
-    showErrorBox({
-      message:
-        i18n._(t`Unable to fetch the example.`) +
-        ' ' +
-        i18n._(t`Verify your internet connection or try again later.`),
-      rawError: error,
-      errorId: 'browser-example-load-error',
-    });
-  } finally {
-    isOpeningCallback(false);
-  }
-};
+import { onCreateFromExampleShortHeader } from './ProjectCreation/services/BrowserCreation';
 
 export const create = (authentication: Authentication) => {
   Window.setUpContextMenu();
